Fix permission issues when copying final image
Since the entire build container runs unprivileged, issues arise when
trying to copy the final image to a bound-mounted path which belongs
to `root` on the host system. Thus, this allows `/bin/cp` to run as
root using a `NOPASSWD` `sudoers` entry.

Unfortunately there is no way to set the `setuid` flag on `/bin/cp`
as it is symlinked to `/bin/busybox`.