Rework QEMU container: use in-house binfmt script, do not run on ARM
This reworks the way the QEMU container works. It does not patch the
script from the original Git repo of QEMU, but it uses a custom version.

This has the advantage of basically eliminating the possibilities of
interfering with pre-existing installations of QEMU on an host, since
this commit introduces an unique interpreter name for the new
`binfmt_misc` entry. In addition, this should also handle the case
where the containers are executed on AArch64 (even though they probably
won't run as the images themselves are x86_64).