Migrate QEMU package from Debian sources to Fedora
The Debian package was fetched directly from the mirrors, which obviously
only retain the latest version of it. Unfortunately this means that they
will irremediably break when an updated package is pushed.

This makes it so that the package is fetched from Fedora's `kojipkgs`
which looks like a reliable and stable source for fixed package versions.

Due to this, the builder now uses `bsdtar` to extract the contents of the
`rpm` file, and drops the dependency of `dpkg`.

Fixes #2.