Build images manually if using docker-compose >= 2.0.0
docker-compose v2 was rewritten from scratch, and as such it introduced regressions.
Particularly, when a local image is expected to be built and another local docker-compose service
refers to it, docker-compose tries to pull the image from the network, failing:
https://github.com/docker/compose/issues/8723

A simple workaround is to pre-build the containers before running `docker-compose up`, so that
Compose is able to find them without failing.

Fixes #29.